grok_client.py
from langchain.llms.base import LLM
import os
from dotenv import load_dotenv
load_dotenv()

class GrokLLM(LLM):
    def _call(self, prompt, stop=None):
        # Simulated Grok API logic (replace with real call)
        return f"Grok response to: {prompt}"

    @property
    def _llm_type(self):
        return "grok"