requirements.txt
langchain
chainlit
xai-grok
torch>=2.0.0
transformers
diffusers
python-dotenv
pillow